+++ miden-lib/src/tests/mod.rs
@@ -8,7 +8,7 @@ use miden_objects::assembly::Library;
 
 #[test]
 fn test_compile() {
-    let path = "miden::kernels::tx::memory::get_consumed_note_ptr";
+    let path = "miden::kernels::tx::memory::get_input_note_ptr";
     let miden = super::MidenLib::default();
     let exists = miden.modules().any(|module| {
         module