+++ miden-tx/src/tests/kernel_tests/test_note.rs
@@ -1,6 +1,6 @@
 use alloc::{collections::BTreeMap, string::String};
 
-use miden_lib::transaction::memory::CURRENT_CONSUMED_NOTE_PTR;
+use miden_lib::transaction::memory::CURRENT_INPUT_NOTE_PTR;
 use miden_objects::{
     accounts::account_id::testing::ACCOUNT_ID_REGULAR_ACCOUNT_UPDATABLE_CODE_OFF_CHAIN,
     notes::Note, testing::prepare_word, transaction::TransactionArgs, Hasher, WORD_SIZE,
@@ -10,7 +10,7 @@ use vm_processor::{ProcessState, EMPTY_WORD, ONE};
 use super::{Felt, Process, ZERO};
 use crate::{
     testing::{
-        utils::consumed_note_data_ptr, MockHost, TransactionContext, TransactionContextBuilder,
+        utils::input_note_data_ptr, MockHost, TransactionContext, TransactionContextBuilder,
     },
     tests::kernel_tests::read_root_mem_value,
 };
@@ -31,8 +31,8 @@ fn test_get_sender_no_sender() {
         begin
             exec.prologue::prepare_transaction
 
-            # force the current consumed note pointer to 0
-            push.0 exec.memory::set_current_consumed_note_ptr
+            # force the current input note pointer to 0
+            push.0 exec.memory::set_current_input_note_ptr
 
             # get the sender
             exec.note::get_sender
@@ -84,7 +84,7 @@ fn test_get_vault_data() {
 
     let notes = tx_context.input_notes();
 
-    // calling get_vault_info should return vault info
+    // calling get_assets_info should return assets info
     let code = format!(
         "
         use.miden::kernels::tx::prologue
@@ -96,23 +96,23 @@ fn test_get_vault_data() {
             # prepare note 0
             exec.note::prepare_note
 
-            # get the vault data
-            exec.note::get_vault_info
+            # get the assets info
+            exec.note::get_assets_info
 
-            # assert the vault data is correct
+            # assert the assets data is correct
             push.{note_0_asset_hash} assert_eqw
             push.{note_0_num_assets} assert_eq
 
-            # increment current consumed note pointer
-            exec.note::increment_current_consumed_note_ptr
+            # increment current input note pointer
+            exec.note::increment_current_input_note_ptr
 
             # prepare note 1
             exec.note::prepare_note
 
-            # get the vault data
-            exec.note::get_vault_info
+            # get the assets data
+            exec.note::get_assets_info
 
-            # assert the vault data is correct
+            # assert the assets data is correct
             push.{note_1_asset_hash} assert_eqw
             push.{note_1_num_assets} assert_eq
         end
@@ -216,8 +216,8 @@ fn test_get_assets() {
             # process note 0
             call.process_note_0
 
-            # increment current consumed note pointer
-            exec.note_internal::increment_current_consumed_note_ptr
+            # increment current input note pointer
+            exec.note_internal::increment_current_input_note_ptr
 
             # prepare note 1
             exec.note_internal::prepare_note
@@ -355,9 +355,9 @@ fn test_note_script_and_note_args() {
 
         begin
             exec.prologue::prepare_transaction
-            exec.memory::get_total_num_consumed_notes push.2 assert_eq
+            exec.memory::get_num_input_notes push.2 assert_eq
             exec.note::prepare_note dropw
-            exec.note::increment_current_consumed_note_ptr drop
+            exec.note::increment_current_input_note_ptr drop
             exec.note::prepare_note dropw
         end
         ";
@@ -393,8 +393,8 @@ fn note_setup_stack_assertions(process: &Process<MockHost>, inputs: &Transaction
 fn note_setup_memory_assertions(process: &Process<MockHost>) {
     // assert that the correct pointer is stored in bookkeeping memory
     assert_eq!(
-        read_root_mem_value(process, CURRENT_CONSUMED_NOTE_PTR)[0],
-        Felt::from(consumed_note_data_ptr(0))
+        read_root_mem_value(process, CURRENT_INPUT_NOTE_PTR)[0],
+        Felt::from(input_note_data_ptr(0))
     );
 }